unittest: Don't throw exceptions from noexcept code
Assert() throws an exception if it fails, so we can't use that.  Instead
set up macros to set a variable and return if an assertion fails in a
unittest testcase, and check that variable in the harness.

(cherry picked from commit d97302bae603885d5cd06e12c1be22fb65691a30)