Fix incorrect results due to uninitialised memory
The array holding max weight values in MultiAndPostList is never
initialised if the operator is unweighted, but the values are still
used to calculate the max weight to pass to subqueries, leading to
incorrect results.  This can be observed with an OR under an unweighted
AND (e.g. OR under AND on the right side of AND_NOT).

The fix applied is to simply default initialise this array, which
should lead to a max weight of 0.0 being passed on to subqueries.

Bug reported in notmuch by Kirill A. Shutemov, and forwarded by
David Bremner.

(cherry picked from commit fa12a83957e97349aa6e2a6c0896faf210dfe4b4)
(cherry picked from commit 7b3349f91ae8e35592fe37606c4443b80577b924)