Allow building with GCC ≥ 7
GCC 7.x enabled the 'implicit-fallthrough' warning when compiling C++,
and Xapian builds with a blanket -Werror when enabling maintainer mode.
This tripped up an implicit fallthrough in a switch block that does not
fall under the "whitelist" approach based on parsing the comments. For
these cases, GCC offers an escape hatch in the form of a compiler
attribute.